refactor: convert-header-panel-file-to-typescript-and-add-type-checks
